+++ packages/html/stories/DragSource.stories.ts
@@ -17,7 +17,6 @@ limitations under the License.
 
 import {
   Graph,
-  type CellState,
   domUtils,
   RubberBandHandler,
   DragSource,
@@ -30,7 +29,6 @@ import {
   CellEditorHandler,
   SelectionCellsHandler,
   ConnectionHandler,
-  type EdgeStyleFunction,
   type DropHandler,
   InternalEvent,
   PanningHandler,
@@ -100,12 +98,6 @@ const Template = ({ label, ...args }: Record<string, string>) => {
       this.options.foldingEnabled = false;
       this.recursiveResize = true;
     }
-
-    override createEdgeHandler(state: CellState, edgeStyle: EdgeStyleFunction | null) {
-      const edgeHandler = super.createEdgeHandler(state, edgeStyle);
-      edgeHandler.snapToTerminals = true; // Enables snapping waypoints to terminals
-      return edgeHandler;
-    }
   }
 
   const graphs: Graph[] = [];
@@ -118,6 +110,19 @@ const Template = ({ label, ...args }: Record<string, string>) => {
 
     const graph = new MyCustomGraph(subContainer);
     graph.gridSize = 30;
+
+    const selectionCellsHandler = graph.getPlugin<SelectionCellsHandler>(
+      'SelectionCellsHandler'
+    )!;
+    // Overrides the createEdgeHandler method to enable snapping waypoints to terminals
+    // Alternatively, you can also subclass SelectionCellsHandler and override the createEdgeHandler method there.
+    // Then register your subclass as a plugin instead of registering the default SelectionCellsHandler plugin.
+    const originalCreateEdgeHandler = selectionCellsHandler.createEdgeHandler;
+    selectionCellsHandler.createEdgeHandler = function (state, edgeStyle) {
+      const edgeHandler = originalCreateEdgeHandler.call(this, state, edgeStyle);
+      edgeHandler.snapToTerminals = true; // Enables snapping waypoints to terminals
+      return edgeHandler;
+    };
 
     // Enables panning
     graph.setPanning(true);